GptIfEngine/requirements.txt
sqlmodel==0.0.8
psycopg2-binary==2.9.6
llama-cpp-python==0.1.33
click==8.1.3
nltk==3.8.1
spacy==3.5.1
https://github.com/explosion/spacy-models/releases/download/en_core_web_sm-3.5.0/en_core_web_sm-3.5.0-py3-none-any.whl
#https://github.com/explosion/spacy-models/releases/download/en_core_web_trf-3.5.0/en_core_web_trf-3.5.0-py3-none-any.whl
rich==13.3.4
PyYAML==6.0
md2py==0.0.1
#torch==2.0.0
fastapi[all]==0.95.0
python-dotenv==1.0.0
requests==2.28.1
zappa==0.56.1
mangum==0.17.0

mypy
black
